refactor: restructure app layout by moving dashboard and profile pages under (main) directory an add layout routing
