+++ kaz-docs/src/app/(main)/dashboard/data.json
