+++ kaz-docs/src/app/(main)/profile/page.tsx
@@ -1,0 +1,11 @@
+import ProfileHeader from "@/components/profile/profile-header"
+import ProfileContent from "@/components/profile/ProfileContent"
+
+export default function Page() {
+  return (
+    <>
+      <ProfileHeader/>
+      <ProfileContent/>
+    </>
+  )
+}
